PE-42104: Add unit tests and fix edge case
This commit adds unit tests for the new jobstat command, and addresses
an unlikely edge case, which is when the API response included a valid
response, did not return an error code, but still returned
'errormessage' and 'errorcode' keys whose values are null.

This commit also adds a TODO comment around the extra error handling
in get_all, which can potentially be removed.

Test Description: ran unit tests