CASMTRIAGE-4288: Fix bootprep errors in copying image
This commit changes the method of copying from using boto3
S3.Object.copy_from to using boto3 s3.Object.copy, which handles
large files better.

Additionally, this change imports botocore.exceptions.ClientError
and catches it in the same places that Boto3Error is caught, so that
future failures will be logged appropriately rather than failing with
a traceback. As a result of importing this exception, botocore is
added to the list of top-level dependencies in requirements.txt.

This commit also removes the HTTP timeout from the IMSClient that
is used in `sat bootprep`, due to IMS taking an unexpectedly long time
to return DELETE requests for large images.

Finally, this commit adds change log entries for these changes.

Test Description: After building this image in Jenkins, asked
an admin to re-run `sat bootprep` with the value of SAT_IMAGE set
to the resulting build.