CRAYSAT-1611: Address KeyError in BOS warning logging
This commit addresses a KeyError when logging a warning that an image
could not be found for a particular BOS component. This warning only
affects the case when both BOS v2 is used and an image does not exist
for a particular BOS component.

This commit also adds a change log entry.

Test Description:
While authenticated to Shandy, I ran `sat status` without this change,
which reproduced the issue. Then, I ran `sat status` with this change,
which addressed the issue.