CRAYSAT-1611: Fix unit test for create_session
One of the SAT unit tests was failing when a local configuration file
existed that specified to use BOS v2 under bos.api_version. This commit
changes the setup of the test so that get_config_value is mocked out,
so that the existence of a local configuration file does not affect the
results of unit tests.

This commit also changes the name of the test_create_session function to
test_create_session_bos_v1 so that it is clear it is only testing BOS
v1. This commit adds a reference to CRAYSAT-1612, a future task to make
sure that the BOS v2 path is covered by unit tests.

Test Description: Ran unit tests.