CRAYSAT-1611: Cache responses from IMS and BOS
This commit adds caching of responses from IMS and BOS to reduce the overall
number of requests when using BOSStatusModule.

In order to get the IMS image name (i.e. the last booted image), a
request is made using the IMS image ID returned by BOS. Prior to this
commit, this resulted in one request per component to get the name
associated with the given IMS image ID. This change introduces a cache
so that we only need to make one request per IMS image ID.

Similarly, a request is made to BOS to get the session associated with
a given component, in order to retrieve the session and session template
names associated with that component. This commit also caches those
responses from BOS, so that a request only needs to be made once per
BOS session ID, rather than once per component.

This commit adds a separate change log entry for this change.

Test Description: I ran `sat status` on shandy with and without this
change, and observed that the command was significantly faster with this
change.