CRAYSAT-1616: Fix SAT cloning CSM repo to get kubectl version
The build script `docker_scripts/config-docker-sat.sh` was failing because
it was attempting to clone a repository that was assumed to be public
over HTTP, in order to get the correct version of `kubectl` to put in
the SAT image.

This commit changes the script to clone a different repository (which has
remained public) to get that information.

Test Description:
* I built the image locally and in Jenkins.
* I verified the version of `kubectl` in `csm-rpms` (new repository) matches
  the version of `kubectl` currently in `node-images` (old
  repository).