CRAYSAT-1633: Make missing kubectl version a fatal error
SAT determines the version of kubectl to use in the cray-sat container
by looking at the version that CSM uses. In particular, it looks at the
csm-rpms repository to see. Prior to this commit, in the event that
the build fails to read the version from csm-rpms, the latest stable
build from Google is used. This may result in less predictable builds
and potentially cause bugs in cases where SAT is reading CLI output
that could be subject to change.

This commit makes the failure case into a fatal error, which leads to
more predictable builds.

Test Description: I built this commit and it failed as expected, due
to a changed path in csm-rpms.