CRAYSAT-1278: Refactor InputConfiguration as subclass of BaseInputItem
In order to reduce code duplication and generalize the behavior across
the different items created by `sat bootprep`, refactor the
`InputConfiguration` as a subclass of the `BaseInputItem` and create an
`InputConfigurationCollection` class which is a subclass of the
`BaseInputItemCollection`.

This allows us to remove the code specific to creating configurations in
the `create_configurations` function, and allows the CFS configurations
to use the common `handle_existing_items`, validation, and
`create_items` methods of the `BaseInputItemCollection`

As part of this change, pass the value of the `--dry-run` option to the
`InputInstance`, so that any item in the instance has access to whether
a dry run is being performed.  As a result, remove `dry_run` parameter
from calls to `handle_existing_items` and `validate` methods of the
`InputSessionTemplateCollection` and `InputConfigurationCollection`
objects.

This could still use more thorough unit tests.

Test Description:
Existing unit tests are passing. Tested on frigg with and without
`--dry-run`. Tested with creating configurations that did not exist yet
and with configurations that did exist already. Tested behavior of
prompt to skip or overwrite images and tested behavior of
`--skip-existing-configs` and `--overwrite-configs` options. Tested
dumping requests to files with the `--output-dir` and `--save-files` in
combination with `--dry-run`.

Tested with some bad CFS configurations to ensure error handling of bad
configuration layers and bad Jinja renders of the `name` field are
handled correctly.