CRAYSAT-1319: Report which items were created by sat bootprep
This change adds a reporting capability to `sat bootprep` which reports
which items (configs, images, and session templates) were created during
a sat bootprep run. This information is printed as a report and its
format can be set to 'pretty', 'yaml' or 'json' based on the context of
the run. (For instance, JSON output is useful for IUF runs, where a
pretty report is more useful for end-user runs.)

This change also adds tests for the new functionality in the
BaseInputItem class.

It should be noted that there may be additional work required to remove
code duplication once the `BaseInputImage` class becomes a subclass of
the `BaseInputItem` class.

Test Description:
Run unit tests. Run built SAT image on internal system with minimal
bootprep input file and check that built items are reported correctly.