CRAYSAT-1606: Implement `--limit` option to `sat bootprep run`
Implement a `--limit` option that allows the user to limit which types
of items from the input file are created. For example, the admin can
choose to create only the CFS configurations or only the IMS images and
BOS session templates.

Skip validation of items when possible, so that validation errors for a
type of item that is not being created do not get in the way of creating
items of the other types. Validation of images can be skipped if neither
images nor session templates are being created. Validation of session
templates can be skipped if session templates are not being created.

This functionality is needed by the IUF because the stages are split
into `update_cfs_config` which will create the CFS configurations and
`prepare_images` which will create and configure IMS images and create
BOS session templates.

Also created constants for the keys in the input file and used it when
reading those properties from the input file and as the possible values
for the `--limit` option.

Test Description:
Tested on frigg with `--limit configurations`, `--limit images`, and
`--limit session_templates` all individually with and without dry-run.
Tested with `--limit images --limit session_templates` combined as well
to ensure the prepare-images IUF stage will work as expected.