CRAYSAT-1319: Refactor Report class, add MultiReport class
This change refactors the Report class to make it easier to dump JSON
and YAML reports. This change also adds a new MultiReport class which
acts as a container for multiple Reports such that they have consistent
formatting and display options. This can also be reused for other
applications where multiple reports are printed such as in `sat
showrev`.

Test Description:
Run unit tests. Test with sample bootprep run on internal system to
check the reporting of created items.