CRAYSAT-1278: Update bootprep man page and usage info
Add the missing `--dry-run` option to the man page, and clarify the
description of that option. Also correct the statement in the
`--save-files` option description that says that only CFS configuration
payloads will be saved. The payloads for the BOS session templates will
now be saved in dry-run mode as well. Add a note that they may be
incomplete in dry-run mode.

Test Description:
Ran `sat bootprep run --help` and viewed usage text. Built man page and
viewed in `man`.