CRAYSAT-1648: Render playbook in CFS config layers in bootprep
Add the ability to render the playbook property of each CFS
configuration layer in the bootprep input file using Jinja2. This will
allow for variable substitution in playbooks, which is useful to support
the two alternative playbooks for the SHS CFS configuration layer.

Add unit tests to check whether this renders correctly.

Test Description:
Unit tests and pycodestyle pass. Tested creating a CFS configuration
with `sat bootprep` on frigg using an input file that used Jinja2
templating in the `playbook` field of a CFS configuration layer.