CRAYSAT-1647: Pass FAS snapshot payload with json parameter
As a result of refactoring SAT to use the common CSM python client,
`sat firmware` changed the data it was passing to the FASClient post()
method, from passing a serialized JSON object to passing a Python
dictionary of the data.

To account for this change, the post() method should take the payload
data with the json parameter rather than the payload parameter.

This commit also updates the relevant unit tests, and adds a change log
entry.

Test Description: I ran `sat firmware -x <xname>` locally with my
development machine targeting shandy. After reproducing the error,
I verified that the commit fixes the issue.