CRAYSAT-1659: Add ability to filter product images/recipes by prefix
If a product provides more than one image or recipe, it was previously
impossible to specify any image or recipe provided by that product. Now
it is possible for the admin to differentiate which image they intended
by specifying a prefix which will be matched against the name of the
image or recipes delivered by that product.

This is needed to support selecting between the Kubernetes image for
masters and workers and the storage image for storage nodes, both of
which are delivered by CSM and can be differentiated by a prefix.

Test Description:
Added unit tests. Tested on frigg with manually populated product
catalog data for the CSM product.