CRAYSAT-1653: Replace hyphens with underscores in top-level keys
This commit changes how bootprep variables are processed. This applies
to variables defined in the software recipe, variables provided on
the command line, and variables in local files (i.e. passed in with
--vars-file).

This changes top-level keys of dictionaries to replace hyphens with
underscores. This gives us the ability to support product names that
contain hyphens. To refer to a product name that contains hyphens,
the bootprep input file will need to refer to that product with
its hyphens replaced with underscores.

Test Description: I ran sat bootprep using an input file that creates
a CFS configuration. Then, using `--save-files` I inspected the
resulting configuration that was created locally.

With a vars file containing a product with a hyphenated name:

    sat bootprep run --dry-run --vars-file vars.yaml \
        configs-hyphen-vars.yaml --save-files

With equivalent variables specified on the command line:

    sat bootprep run --dry-run \
        --vars "slingshot-host-software.version=2.0.0-75" \
        configs-hyphen-vars.yaml --save-files