CRAYSAT-1657: Add --staged-session
This commit adds a new CLI argument, --staged-session which can only be
used with BOS v2. When specified, the 'stage' field will be set to True
in BOS session creation. This creates a session but does not immediately
change the state of targeted components. Instead, those components'
get their "staged_state" updated. Later, a separate call to the BOS API
or `cray` CLI can change the staged state to a "desired_state", which
will start the process of nodes booting, rebooting or shutting down.

This commit modifies the do_bos_operations method to accept a 'stage'
parameter. When specified, the command will create 'staged' BOS sessions
and the command will not wait for their completion, but instead wait for
the sessions to reach the "running" state, indicating that the
components' "staged_state" has been updated.

As a part of this change, when creating a staged session, the command
bypasses the check on whether templates need an operation. In other
words, creating a staged session from a template always creates a
session, whereas normal sessions will not be created if components do
not need the specified operation performed.

This commit also updates the `sat bootsys` man page, including adding a
couple options that were previously missing, and adds a change log
entry.

Test Description:
* Ran the following commands to test the behavior of `--staged-session`:

        sat bootsys --bos-version=v2 boot --bos-templates <template> \
                --stage bos-operations --bos-limit <xname> --staged-session

        sat bootsys --bos-version=v2 reboot --bos-templates <template> \
                --stage bos-operations --bos-limit <xname> --staged-session

        sat bootsys --bos-version=v2 shutdown --bos-templates <template> \
                --stage bos-operations --bos-limit <xname> --staged-session

* Ran the following commands to test the behavior without
  `--staged-session`:

        sat bootsys shutdown --bos-templates <template> --bos-limit <xname> \
                --stage bos-operations
        sat bootsys boot --bos-templates <template> --bos-limit <xname> \
                --stage bos-operations
        sat bootsys reboot --bos-templates <template> --bos-limit <xname> \
                --stage bos-operations

* Ran the following to ensure BOS V1 is not allowed with
  `--staged-session`:

        sat bootsys --bos-version=v1 reboot --bos-templates <template> \
                --bos-limit <xname> --staged-session \
                --stage bos-operations