CRAYSAT-1604: Add configurable retries
This commit adds two new options, `--api-retries` and `--api-backoff`
that control the number of retries and interval between retries. These
options can be specified via the config file or on the command line.

To implement these options, the urllib3.util.Retry class is used, which
is passed to the requests.Session using an HTTPAdapter class. Retries
will only be done for status codes in the 500 range or for network
errors, errors like "400 Bad Request" or "403 Forbidden" will not
result in any retries.

To ensure that retries are logged, handle log messages from the
`urllib3` module.

This commit also documents the new command-line and config file options
in the main sat(8) man page.

Add a change log entry in a new "Unreleased" section.

Test Description:
* I ran `sat status` locally with an authenticated session to ensure
  normal usage was working.
* I temporarily moved my authentication token aside, resulting in
  401 Unauthorized errors from the system when running `sat status`.
  I verified that it was not possible to add retries in this case.
* I modified my configuration file to specify an incorrect URL for
  the system's API gateway, resulting in network-level errors when
  running `sat status`. I verified that I was able to change both
  the number of retries and backoff factor both by using command-line
  options as well as configuration file options.
* I ran `sat bootprep` to create images, configurations and session
  templates on an internal system.