CRAYSAT-1656: Use updated load_kube_api() implementation
Bump the version of the csm-api-client dependency to 1.1.4 in order to
prioritize loading the Kubernetes config from the in-cluster config, and
to fall back to the config file if not running in the K8s cluster.

This change also utilizes the newer implementation of the
load_kube_api() function in all instances where the Kubernetes API is
being accessed in order to reduce duplication and make the Kubernetes
configuration loading more consistent across subcommands.

Test Description:

Run in k8s cluster with DEBUG logging, ensure no message logged. Check
running directly on management NCN and make sure DEBUG message was
logged.