CRAYSAT-1656: Bump csm-api-client version to 1.1.3
Bump the version of the csm-api-client dependency to 1.1.3 in order to
prioritize loading the Kubernetes config from the in-cluster config, and
to fall back to the config file if not running in the K8s cluster.

Test Description:

Run in k8s cluster with DEBUG logging, ensure no message logged. Check
running directly on management NCN and make sure DEBUG message was
logged.