CRAYSAT-1523: Migrate unit tests to nose2 runner
This change modifies this repo's testing infrastructure in order to use
nose2 for running unit tests instead of nosetests. This should
future-proof our unit tests as nosetests breaks after Python 3.10 and
will not be updated to support it (the nose project is in maintenance
mode.)

Test Description:
Run unit tests. Build testing image in CI pipeline and check that it
passes.