CRAYSAT-1705: Fix logic for checking when staged BOS session is complete
Normally, a BOS v2 staged session will reach the state "running" when it
is considered finished staging. However, if that BOS v2 staged session
does not apply to any components (for example, when a limit parameter is
used that does not overlap with any of the nodes in the session
template), the session will go straight to "complete" instead.

Rename the `target_state` argument to the BOSV2SessionWaiter to
`target_states` and change the type to a list of acceptable target
states. This allows us to call it with the single state "complete" when
waiting on a non-staged session and to call it with the states
"complete" and "running" when waiting on a staged session.

Test Description:
Added new unit tests. Will test on a real system as follows:

* Run `sat bootsys boot --stage bos-operations --staged-session` using
  a session template that applies to application nodes, but pass
  `--bos-limit Compute`. The session should finish and `sat bootsys`
  should return.
* Run the same test as above but pass two session templates: one for
  compute nodes and one for UANs. Both should finish and `sat bootsys`
  should return. Only computes should be staged.
* Run the same test as above but without `--staged-session`. The
  computes should actually be rebooted.

(cherry picked from commit ff6787011b073fc3493a585254c011011dce05c1)