CRAYSAT-1673: Update csm-api-client version
Use the 1.1.3 version of the csm-api-client dependency in order to pull
in the fixed suggested debug container logs functionality to accomodate
new versions of CFS with only one ansible container.

Test Description:
Test with intentionally failing bootprep image configuration session and
confirm that the `ansible` container is returned instead of the
`teardown` container.