CRAYSAT-1714: Skip CAPMC request when empty set of nodes
During `sat bootsys shutdown --stage cabinet-power`, skip making a
request to CAPMC and waiting on the power state of nodes when there are
no non-management nodes in air-cooled cabinets to ensure are powered
off. Instead, simply log an info message to that effect. Before this
check, we were making a request to CAPMC to power off an empty set of
xnames, which resulted in a confusing warning, and then we were waiting
on a set of zero components, which also looked confusing.

Add unit tests for `do_air_cooled_cabinets_power_off`.

Test Description:
Unit tests should be sufficient here until we can do a test on a real or
vshasta system.