CRAYSAT-1719: Update kubectl version lookup
This change updates the process used to retrieve the appropriate version
of kubectl to download into the container. A Python script is now used
to parse out the semantic version of kubectl listed in the
metal-provision repository instead of the deprecated csm-rpms
repository.

Test Description:
Build cray-sat container image. Check kubectl version in image.