CRAYSAT-654: Remove cronjob recreate step from cabinet-power stage
This change removes recreating all "stuck" cronjobs from the
cabinet-power stage since there was specific functionality added to
recreate the hms-discovery cronjob in that stage, which would cause the
cronjob to be recreated unconditionally. This removes this recreating
step from cabinet-power so that the bootsys behavior is in line with
docs-csm.

Test Description:
Run unit tests.