CRAYSAT-1718: Implement product image filter by arch
Implement the ability to filter images or recipes provided by product
based on their architecture in the IMS resource object.

It is also now possible to specify more than one filter in the
`base.product.filter` object in the image definitions in the bootprep
input file. E.g., you can specify both a prefix and an architecture.

Note that the filtering logic had to be moved to occur after the
resources are obtained from IMS because we are filtering on more than
just the name of the IMS image, which was available in the product
catalog. Now we are also filtering on another property (`arch`) of IMS
images and recipes. This can result in potentially a few more requests
to IMS when a product provides multiple IMS images or recipes, but it
seemed worth it to allow us to do filtering all at once in a consistent
manner rather than once on the reslts from the product catalog and again
with additional information from IMS.

Modify and add new unit tests accordingly.

This needs corresponding documentation updates to the bootprep usage
section.

Test Description:
Unit tests and pycodestyle passing.

Not tested on a real system yet, but will use mug to do so when
available.