CRAYSAT-1733: Adjust bootprep multi-arch feature to match BOS
Adjust the implementation of `sat bootprep` support for multiple
architectures to match the implementation of BOS. This includes:

* Remove the ability to specify the `arch` at the top level of the BOS
  session template. BOS will not support this for now.
* Change the guidance in the `sat bootprep` input file schema to say
  that the BOS session template architecture values should match what
  HSM returns for its components API endpoint.
* Add an enum for validation of the `arch` field in a boot set in a BOS
  session template to provide clear guidance to the user when an invalid
  value is used.
* Update the description of BOS's default behavior when no value is
  specified for the `arch` field. It now treats this as "X86".

Update the unit tests accordingly for the above changes.

Note that technically, this is a backwards-incompatible schema change,
and the bootprep input file schema's version would normally need to be
updated accordingly. However, that previous version was not part of a
stable sat build yet, so it's preferable not to unnecessarily bump the
schema major version and unnecessarily cause issues for existing input
files.

Test Description:
Unit tests pass. Tested creation of a BOS session template in dry-run
mode on mug.

Built bootprep input file schema documentation with `sat bootprep
generate-docs` and viewed in browser.