CRAYSAT-1733: Add more detail to unit test failure messages
Add more detail when the `assert_regex_matches_element` method of the
`ExtendedTestCase` class fails. Include the elements in the container,
so that we can see which elements are actually present and being
compared against the regular expression.

This was useful in debugging an incorrect unit test in
`tests.cli.bootprep.test_validate`.

Test Description:
All unit tests pass.

Tested with an intentional failure of a unit test in `test_validate` and
got output like the following:

```
AssertionError: Regex (    \['images'\]\[0\]\['bos_parameters'\]\['boot_sets'\]\['compute'\]\['arch'\]:.*'aarch64' is not one of) does not match any of the string elements in the given container:
Input file is invalid with the following validation errors:
    ['session_templates'][0]['bos_parameters']['boot_sets']['compute']['arch']: 'aarch64' is not one of ['X86', 'ARM', 'Other', 'Unknown']
```