CRAYSAT-1727: Simplify bos-operations stage of `sat bootsys`
Simplify the bos-operations stage of `sat bootsys`. Instead of
determining whether the given session templates need operations to be
performed against them, just request that BOS perform the given
operation (boot, shutdown, or reboot) against them, and let BOS
determine whether the operation is needed. With the addition of an
`arch` field in boot sets in BOS session templates, more logic would
have been needed to correctly identify nodes affected by a session
template, and this logic duplicates existing logic in BOS.

This code was originally an attempt to achieve idempotency in `sat
bootsys` so that BOS sessions are not needlessly created. However, BOS
already handles these cases well.

In BOS v1, if the operation is `boot`, and the nodes are already up, BOS
will not attempt to reboot them.

In BOS v2, if the operation is `boot`, and the nodes are already up, BOS
will check whether they are booted with the correct artifacts, and if
not, it will reboot them.

In both BOS v1 and v2, if the operation is `shutdown`, and the nodes are
already off, BOS will not do anything to them. If the operation is
`reboot`, the nodes are always rebooted.

Test Description:
Unit tests pass.

Tested with `sat bootsys boot --stage bos-operations` and `sat bootsys
shutdown --stage bos-operations` commands on odin. Ensured that it
behaves correctly correctly whether the nodes are already in the desired
state or not and with both BOS v1 and BOS v2.