CRAYSAT-1750: Update PyYAML to 6.0.1
Update PyYAML to 6.0.1 to fix a build error. This build error was due to
an incompatible change introduced in Cython 3.0. The PyYAML package has
worked around this in version 6.0.1 by restricting to Cython<3.0.

Also update the pinned version of csm-api-client to 1.1.5, which relaxed
the PyYAML constraint to allow versions greater than or equal to 6.0.
Without this change, `pip` is unable to resolve sat's dependencies.

Test Description:
Ran the following `sat` commands to exercise relevant code:

* `sat showrev --products` (config map access)
* `sat bootprep run` (YAML loading, API access, config map access)
* `sat status` (multiple API access)