CRAYSAT-1691: Limit CFS config name length to 63 chars in bootprep
Limit the name of the CFS configurations which can be created with `sat
bootprep` to 63 characters, which is the limit for Kubernetes labels. In
CFS v2, if a CFS configuration with a name longer than this is used in
an image customization session, the CFS session fails in a mysterious
way. In CFS v3, such CFS configuration names are not allowed.

Enforcing this limit in `sat bootprep` should result in less confusion
for admins.

Test Description:
New unit tests added. Executed `sat bootprep` on mug with valid and
invalid configuration names, both in dry-run and regular mode.