CRAYSAT-1780: Retain image arch in `sat bootprep`
When renaming images in `sat bootprep`, ensure the image architecture is
retained in the new image. This field is important to IMS, so it knows
how to handle any future customization operations on that image.

Test Description:
Tested on mug by downloading the new cray-sat image, opening a shell in
the container with `sat bash`, and then opening a Python interpreter and
using the `copy_image` function directly to confirm that the image
architecture is retained when copying an aarch64 image. Also tested on
an x86_64 image for good measure.