CRAYSAT-1789: Allow `additional_inventory` in bootprep CFS configs
Add support to the bootprep input file schema to support specifying
`additional_inventory` when creating CFS configurations using `sat
bootprep`.

This includes a schema change for the bootprep input file that allows
the user to specify a new `additional_inventory` property in their CFS
configuration definitions which corresponds with the field of the same
name in the CFS API.

Add a unit test that just verifies that the `bootprep_schema.yaml` data
file is valid against the JSON Schema metaschema. This is helpful to
test only the validity of edits to the bootprep input file schema.

Add unit tests to validate against the schema new bootprep input that
includes the new `additional_inventory` property.

When `additional_inventory` is specified, treat branches in the same way
as for layers. That is, by default, resolve them to commit hashes before
passing them to CFS. If `--no-resolve-branches` is specified, pass the
branches as-is to CFS and let it do the resolution.

Utilize the same code for branch resolution that is already used for the
layers of a CFS configuration. To do so, split the common functionality
between the layers and the additional_inventory into a new common
abstract base class called `InputConfigurationLayerBase`. Then implement
the remaining abstract methods in `InputConfigurationLayer` and
`AdditionalInventory` subclasses.

Add and update unit tests to test the `AdditionalInventory` subclass.

Test Description:
Unit tests pass.

Tested on mug by creating a bootprep input file that defines CFS
configurations. Defined several CFS configurations including:

* One with no additional_inventory and a layer without a branch
* One with no additional_inventory and a layer with a branch
* One with additional_inventory and layer, both with branches
* One with additional_inventory and layer, both with commits
* One with additional_inventory with an invalid branch name
* One with additional_inventory and layer with invalid branch names

Executed `sat bootprep run` against this input file and confirmed that
CFS configurations were created as expected both with and without the
`--no-resolve-branches` option.