CRAYSAT-1800: Fix logging traceback due to deepcopy
The use of `deepcopy` in the `LineSplittingLogFormatter` causes problems
when a logging method is called, and a format string is passed in along
with an argument to be used in that format string, and the object does
not support copying with `deepcopy`.

Switching to `copy` instead. This is all that is needed because the
`LineSplittingLogFormatter` was just creating a copy of the log record
to modify its message without modifying the original record. Using
`copy` is sufficient for this purpose.

This fixes an issue seen when a filter query string could not be parsed
and a logging call is made that passes along the `ParseError` object
from `parsec`.

Add a unit test that ensures that the original log record's message is
not indavertently changed, which is what the `deepcopy` was originally
trying to ensure.

Test Description:
Unit tests for logging multiple lines with the logging format applied to
each line still work.

Tested `sat bootprep` against an input file that had multiple errors,
and verified that every line of the error message was prefixed with the
"ERROR: " log level.

Tested `sat status --filter 'bad'` and confirmed that it just logs a
simple error message containing the string representation of the
`ParseError` instead of a long traceback.