CRAYSAT-1798: Preserve all port policies in `sat swap`
When enabling or disabling a switch or a cable in `sat swap switch` or
`sat swap cable`, properly handle when a port has multiple port policies
applied to it. This fixes the issue where only the first port policy
associated with a port is kept through the disable/enable cycle.

The strategy for keeping a record of the port policies applied to each
port is unchanged. That is, for the disable action, for each port
policy associated with the ports to be disabled, append the
"sat-offline-" prefix to the policy name, create a new policy
with the prefixed name which has `"state": "OFFLINE"`, and then apply
the policy to the port being disabled.

Note that for ports with more than one policy associated with them (e.g.
fabric ports which have a fabric policy and a qos policy), this will
result in multiple policies with the same `"state": "OFFLINE"` data
being assigned to those ports. According to the Slingshot team, this is
acceptable and should not cause issues.

Fix and improve the unit tests for the `get_switch_port_data_list`
method of the `PortManager`.  This now tests the method for a switch
with just edge ports, a switch with just fabric ports, and a switch with
both fabric and edge ports. It also does more consistent mocking by
mocking the `get_port` method instead of directly mocking the
`FabricControllerClient` class.

Fix unit tests for `PortManager.get_jack_port_data_list`. Fix the
`swap_component` unit tests for the `CableSwapper` and `SwitchSwapper`.

Fix minor issues in tests including PyCharm code inspection type
warnings and faulty `assertLogs` usages.

Test Description:
The following will be tested on baldar:

* `sat swap switch --action disable` on an enabled switch
* `sat swap switch --action disable` on an disabled switch
* `sat swap switch --action enable` on an disabled switch
* `sat swap switch --action enable` on an enabled switch

The same will also be tested for `sat swap cable`.