CRAYSAT-1798: Remove unused `disruptive` argument from `swap_component`
The `disruptive` argument was not used in the `swap_component` method of
the `Swapper` class. The value of the `disruptive` command-line option
is checked in the function `check_arguments` in `sat.cli.swap.main`
before calling the appropriate function for the cable, switch, or blade.

Test Description:
Unit tests pass. PyCharm is happier.