CRAYSAT-1621: Modify shutdown order in platform-services & ncn-power
The changes have be made on two stages of bootsys:

1. sat bootsys shutdown -- platform-services

freeze ceph has been cleaned from the this stage.

2. sat bootsys shutdown -- ncn-power:

Here we have modified the way shutdown of the ncn's are performed.  Earlier all
the ncn's except m001 would be shutdown simultaneously together. Now it would
be performed in grouping. And the order would be workers, masters (except
m001), unmap rbd on m001, freeze ceph and then shutdown the storage nodes.